Web/requirements.txt
django==2.2
mysqlclient==1.3.14
django-cors-headers
djangorestframework
docker-py
rsa
Pillow
pycryptodome
